test(svelte-query): unify queryClient passing to 'setQueryClientContext' and rename test components
