+++ packages/svelte-query/tests/QueryClientProvider/BaseExample.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
   import { QueryClientProvider } from '../../src/index.js'
-  import ChildComponent from './ChildComponent.svelte'
+  import QueryChild from './QueryChild.svelte'
   import type { QueryClient } from '@tanstack/query-core'
 
   let { queryClient }: { queryClient: QueryClient } = $props()
 </script>
 
 <QueryClientProvider client={queryClient}>
-  <ChildComponent />
+  <QueryChild />
 </QueryClientProvider>
