+++ packages/svelte-query/tests/QueryClientProvider/QueryChild.svelte
