+++ packages/svelte-query/tests/createInfiniteQuery/BaseExample.svelte
@@ -3,21 +3,20 @@
   import { QueryClient } from '@tanstack/query-core'
   import { queryKey, sleep } from '@tanstack/query-test-utils'
   import { createInfiniteQuery } from '../../src/index.js'
+  import { setQueryClientContext } from '../../src/context.js'
   import type { QueryObserverResult } from '@tanstack/query-core'
 
   let { states }: { states: { value: Array<QueryObserverResult> } } = $props()
 
   const queryClient = new QueryClient()
+  setQueryClientContext(queryClient)
 
-  const query = createInfiniteQuery(
-    () => ({
-      queryKey: queryKey(),
-      queryFn: ({ pageParam }) => sleep(10).then(() => pageParam),
-      getNextPageParam: (lastPage) => lastPage + 1,
-      initialPageParam: 0,
-    }),
-    () => queryClient,
-  )
+  const query = createInfiniteQuery(() => ({
+    queryKey: queryKey(),
+    queryFn: ({ pageParam }) => sleep(10).then(() => pageParam),
+    getNextPageParam: (lastPage) => lastPage + 1,
+    initialPageParam: 0,
+  }))
 
   $effect(() => {
     // @ts-expect-error
