+++ packages/svelte-query/tests/createInfiniteQuery/ChangeClientExample.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import { QueryClient } from '@tanstack/query-core'
   import { createInfiniteQuery } from '../../src/index.js'
+  import { setQueryClientContext } from '../../src/context.js'
   import { sleep } from '@tanstack/query-test-utils'
 
   let { queryClient }: { queryClient: QueryClient } = $props()
@@ -9,15 +10,14 @@
 
   let firstPage = $state(0)
 
-  const query = createInfiniteQuery(
-    () => ({
-      queryKey: queryKey,
-      queryFn: ({ pageParam }) => sleep(10).then(() => pageParam),
-      getNextPageParam: (lastPage) => lastPage + 1,
-      initialPageParam: firstPage,
-    }),
-    () => queryClient,
-  )
+  setQueryClientContext(queryClient)
+
+  const query = createInfiniteQuery(() => ({
+    queryKey: queryKey,
+    queryFn: ({ pageParam }) => sleep(10).then(() => pageParam),
+    getNextPageParam: (lastPage) => lastPage + 1,
+    initialPageParam: firstPage,
+  }))
 </script>
 
 <button
