--- packages/svelte-query/tests/useIsFetching/FetchStatus.svelte
@@ -1,6 +1,0 @@
-<script lang="ts">
-  import { useIsFetching } from '../../src/index.js'
-  const isFetching = useIsFetching()
-</script>
-
-<div>isFetching: {isFetching.current}</div>
