+++ packages/svelte-query/tests/useIsMutating/BaseExample.svelte
@@ -1,11 +1,20 @@
 <script lang="ts">
-  import ProviderWrapper from '../ProviderWrapper.svelte'
-  import MutatingStatus from './MutatingStatus.svelte'
-  import Query from './Query.svelte'
+  import { QueryClient } from '@tanstack/query-core'
+  import { queryKey, sleep } from '@tanstack/query-test-utils'
+  import { setQueryClientContext } from '../../src/context.js'
+  import { createMutation, useIsMutating } from '../../src/index.js'
+
+  const queryClient = new QueryClient()
+  setQueryClientContext(queryClient)
+
+  const mutation = createMutation(() => ({
+    mutationKey: queryKey(),
+    mutationFn: () => sleep(10).then(() => 'data'),
+  }))
+
+  const isMutating = useIsMutating()
 </script>
 
-<ProviderWrapper>
-  <MutatingStatus />
+<button onclick={() => mutation.mutate()}>Trigger</button>
 
-  <Query />
-</ProviderWrapper>
+<div>isMutating: {isMutating.current}</div>
