--- packages/svelte-query/tests/useIsMutating/MutatingStatus.svelte
@@ -1,6 +1,0 @@
-<script lang="ts">
-  import { useIsMutating } from '../../src/index.js'
-  const isMutating = useIsMutating()
-</script>
-
-<div>isMutating: {isMutating.current}</div>
